fixed dtype for hydrodynamic coefficients
due to the pandas.DataFrame.explode() function in dataframe_from_bemio, all of the hydrodynamic coefficients were having their dtypes recast as objects, which results in a loss of metadata (e.g. no max/min evaluation). Pandas does not currently have a built-in method to infer the dtype from the values when using explode (though there appears to be a feature request for this), so we have to manually do this after the explode() function.