Add import of Bemio data as Xarray datasets (#79)
Bemio class types can now be read and formatted into an xarray.Dataset formatted the same as a LinearPotentialFlowResult in assemble_dataset. 

Bemio does not natively support some of the parameters that Capytaine does, so some of the optional arguments used when assembling the xarray dataset do not apply. If any of these optional arguments are set to True, the program now prints a notice that the option is skipped.

Natively, WAMIT returns nondimensionalized outputs (except for the volume and center of buoyancy), whereas AQWA and NEMOH return dimensionalized values. Therefore, sending Bemio imports to Capytaine could be prone to error. The Bemio importing functionality now dimensionalizes the WAMIT values properly, as per Chapter 4 of the WAMIT manual.

Co-authored-by: Ryan Coe <rcoe@sandia.gov>
Co-authored-by: Matthieu Ancellin <matthieu.ancellin@ens-paris-saclay.fr>